Rename galaciuc component to about
